// kyrvo-frontend/src/models/Web.ts
export interface Web {
    Id?: string;
    SystemName?: string;
    SystemImage?: string;
    SystemPrimary?: string;
    SystemSecondary?: string;
    SystemTertiary?: string;
    IntroTop?: string;
    IntroImage?: string;
    ChiefMessageImage?: string;
    CarFeatureImage?: string;
    CreatedOn?: number;
    UpdatedOn?: number;
}